Remove data processor module
- Deleted the entire data processor module, including clients, handlers, schemas, routers, and services.
- This cleanup eliminates unused code and simplifies the project structure.
