--- modules/data_processor/__init__.py
@@ -1,3 +1,0 @@
-from .routers import router as data_processor_router
-
-__all__ = ["data_processor_router"]
