--- modules/data_processor/dependencies.py
@@ -1,9 +1,0 @@
-# from fastapi import Depends
-# from app.core.adapter import MultiProviderClient
-# from .handlers import DatasetHandler
-
-# def get_dataset_handler():
-#     return DatasetHandler()
-
-# def get_visualization_service(handler: DatasetHandler = Depends(get_dataset_handler)):
-#     return DataVisualizationService(handler)
