+++ modules/projects/dependencies.py
@@ -1,0 +1,4 @@
+from modules.projects.clients import ProjectClient
+
+def get_project_client() -> ProjectClient:
+    return ProjectClient()
